Add version generation logic
Description
===========

This is the first patch with the initial functionality for this plugin.
In essence this is a modernized striped down version of the
[`net.wooga.release`] plugins version generation code. This plugin
provides only one functionality at the moment which is to infer a
opinionated semver version from the current git state.
For this to work I decided to fork some dependencies of
[`net.wooga.release`] and patch them directly. My plan is to add new
custom functionality and remove logic we don't need/want.
I added sources from [gradle-git] and [nebular-release] and adjusted
them directly to remove some linkages. We don't want to backport any
changes from the upstream projects so this is safe.

The plugins hooks into gradle just like [gradle-git] did and sets a
`DelayedVersion` object for all gradle projects in the build. The plugin
can only be applied to the root project but future planned patches will
bring other changes that make it possible to use the plugin also in sub
projects. This is the reason for the non standart setup with a hidden
_base_ plugin which holds all major logic.

I copied the test specs for the wooga semver1 and semver2 strategies as
is from [`net.wooga.release`] to make sure we have a drop in replacement
with this plugin.

Changes
=======

* ![ADD] version generation logic

[`net.wooga.release`]:  https://github.com/wooga/atlas-release
[gradle-git]:		https://github.com/ajoberstar/gradle-git
[nebular-release:	https://github.com/nebula-plugins/nebula-release-plugin/tree/v6.3.5