Refactor version provider (#6)
Description
===========

This is a small cleanup patch to prepare the next `versionCode`
implementation. I renamed the extension property `scheme` to
`versionScheme` and also adjusted the type to an enum `VersionScheme`. I
left the ability to set this value from `String` to make it easier to
set the value in the `build.gradle` script.

I wrapped the version provider in a new custom provider which can cache
the infered value. This whole provider gets wrapped again in a special
`ToStringProvider`. This providers `toString` method will evaluate the
internal value and call `toString` on it. The default `Provider`
instances only provide a basic `toString` implementation.

I also added more test cases for the new APIs and classes

Changes
=======

* ![CHANGE] `scheme` to `versionScheme`
* ![CHANGE] `versionScheme` type from `String` to `VersionScheme`
* ![ADD] `MemoisationProvider`
* ![ADD] `ToStringProvider`
* ![IMPROVE] `versionScheme` API overloads