Add versionCode generation (#7)
Description
===========

This patch adds multiple schemes to generate or manual set a
versionCode. The strategies differ in two major ways:

semver version based
--------------------

A strategy based on a `semver` version number. This patch provides two
`schemes` to generate a version code from a given semver version string:

_`semverBasic`_

This scheme turns a semver version `major.minor.patch` into a integer
representation. Everything after the components is ignored. The version
components are not allowed to be higher than `99` This means that the
biggest version provided possible is `99.99.99`.
The resulting code is an integer with padded values.

`1.0.0`   => `10000`
`1.1.1`   => `10101`
`1.11.10` => `11110`

_`semver`_

The `semver` scheme is an opinionated addon to the `semverBasic` scheme.
The resulting code is extended by a revision counter. This counter has
only one digit so can't be higher than `9`. The revision is parsed from
a version with `-staging.{revision}` appended.

`1.0.0`           => `100000`
`1.0.0-staging.3` => `100003

release count based
-------------------

These schemes act more as a build number generator by counting all
reachable release tags from the current branch. An optional `offset` can
be provided to keep the generated version-code over a certain limit.

_`releaseCountBasic`_

Counts all reachable tags in the repo with final release versions
appended (e.g `v1.1.0`). It will skip any prerelease tags or any other
tags.

_ `releaseCount`_

Counts all kinds of release tags reachable from the current branch.
(e.g. `v1.1.0`, `v1.2.0-staging.1`)

configuration
-------------

*build.gradle*

```groovy
plugins {
    id "net.wooga.version" version "VERSION"
}

versionBuilder {
   versionScheme = "semver2"
   versionCodeScheme = "releaseCount"
   versionCodeOffset = 1000
}
```

You can also manually set a versionCode. This has to be done after this
plugin is applied.

```groovy
plugins {
    id "net.wooga.version" version "VERSION"
}

allprojects { prj ->
   //use the property created by the plugin
   prj.versionCode.set(12345)

   //replace the extension
   prj.ext.versionCode = 12345
}
```

In both cases, the effictive versioncode will be `12345` but be carefull
when replacing the extension. You could change the type of the value.
And the behavior changes. Use the provided `Property.set` method.

The plugin will set a `versionCode` extension to all projects. This
value is of type `VersionCodeExtension` which implements the gradle
`Provider` interface. To retrieve the value you need to call `get` or
`getOrNull` or `getOrDefault` The version-code generation process is
lazy like the version infer logic. If you pass this value to a task that
also uses the provider API, you can use this value directly or call
`map` to convert it to a fitting type.

Changes
=======

* ![ADD] version-code generation