Fix TagStrategy tag prefix parsing default value (#8)
Description
===========

With the patch [#7] I added a regression. Because of a type I disabled
parsing of tags without a prefix `v`. I had no tests in place for this
behavior so this ended up in the latest `rc` build [v0.1.0-rc.4].

My intention was to set the new added field
`parseVersionsWithoutPrefixV` to true for the default constructor.
Instead I picked the groovy generated `field` for the
`setPrefixNameWithV` (`prefixNameWithV`).

I fixed the issue and added a small set of regression tests.

Changes
=======

* ![FIX] `TagStrategy` tag prefix parsing default value

[#7]:  https://github.com/wooga/atlas-version/pull/7
[v0.1.0-rc.4]: https://github.com/wooga/atlas-version/releases/tag/v0.1.0-rc.4