Add new strategy staticMarker (#9)
Description
===========

This new strategy won't auto increment the normal part of the version as
`semver` or `semver2` would do. Instead the whole normal part of the
version is taken from a `marker` tag. Other metadata information for
prerelease versions are generated in a similar fashion than done in the
`semver` strategies but not based on the latest release tag but also on
the marker tag.

One can define two marker tags in the project.

* `ci-0.0.0`
* `release-0.0.0`

The prefix can not be configured at the moment. The version strategy
will look for these markers and determine the normal of the generated
version. We have two different markers depending on the current release
cycle. We can configure to use a specific marker based on the current
git branch. This can be configured with the `releaseBranchPattern`
property which has a default value of `/(^release\/.*|^master$)/`. Which
means if the current branch is named `master` or starts with `release/`
then use the normal from the `release-*` marker tag. If not use the ci
marker. This allows to bump the ci version to the next to be released
version for development when the release process is kicked off in a
sidebranch. To illustrate this:

```
         *     branch: 'develop', version: '0.2.0-develop.2
        /|
       ◊ |     branch: 'master', version: '0.1.0', tag: 'v0.1.0,
       | |
       * |     branch: 'master', version: '0.1.0-master.1, distance: 1
       | |
       | *     branch: 'develop', version: '0.2.0-develop.1
        \|
         ◊     branch: 'develop', version: '0.0.0', tag: 'release-0.1.0', ci-0.2.0,
```

Staging and production build generation works as before. Staging builds
will increment based on the previous staging version.

This patch also adds a new configuration property which allows to
configure which branch(es) are the main development branches. This
information is used to construct the snapshot prerelease metadata.

A snapshot version gets constructed with the normal version part +
encoded branchname + counter.

The branchname comes in two flavors.
If it is a main branch (`master`, `develop`) by default it won't be
prefixed with `branch.` (depending on the semver strategy)
Now it's possible to provide a patter for this check

`mainBranchPattern` the default value is `/(^master$|^develop$)/`

Changes
=======

* ![ADD] `staticMarker` version strategy