Fix jgit jdk8 compatibility issue
Description
===========

This patch updates `grgit-core` to patch `4.1.1` which limits the
version range of `jgit` to latest major version `5`. This allows
this plugin to be used with jdk8 still. I also had to add an
force resolution of `grgit-core:4.1.1` because the buildscript
pulls a version of this plugin into the classpath. This should later
be saved to remove

Changes
=======

* ![FIX] `jgit` jdk8 compatibility issue